+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed 
+
+- Cleanup all annotations.
+
 ## [0.3.1] - 2022-12-02
 
 ### Fixed